+++ packages/expo-module-template-local/snippets/SharedObject/app.imports.tsx.ejs
@@ -1,0 +1,1 @@
+import { use<%- project.sharedObjectName %> } from '<%- project.slug %>';
