+++ packages/expo-module-template/snippets/Event/app.imports.tsx.ejs
@@ -1,0 +1,1 @@
+import { useEvent } from 'expo';
