+++ packages/expo-module-template/snippets/SharedObject/app.imports.tsx.ejs
@@ -1,0 +1,1 @@
+import { use<%- project.sharedObjectName %> } from '<%- project.slug %>';
