+++ packages/expo-module-template/snippets/View/app.imports.tsx.ejs
@@ -1,0 +1,1 @@
+import { <%- project.viewName %> } from '<%- project.slug %>';
